# docs/swagger.yaml
basePath: /api
definitions:
...
  model.AttemptCreateRequest:
    properties:
      chapterId:
        type: integer
      correctAnswer:
        type: string
      goalId:
        type: integer
      isCorrect:
        type: boolean
      maxScore:
        type: integer
      questionContentId:
        type: integer
      questionId:
        type: integer
      score:
        type: integer
      studentAnswer:
        type: string
      taskId:
        type: integer
      timeSpentSeconds:
        type: integer
    required:
    - chapterId
    - goalId
    - questionContentId
    - questionId
    type: object
  model.AttemptQuery:
    properties:
      chapterId:
        type: integer
      dateFrom:
        type: string
      dateTo:
        type: string
      goalId:
        type: integer
      id:
        type: integer
      pageIndex:
        type: integer
      pageSize:
        type: integer
      questionId:
        type: integer
      taskId:
        type: integer
      userId:
        type: integer
    type: object
...
info:
  contact:
    name: LTEdu Support
  description: LTEdu 在线教育平台后端接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LTEdu API
  version: "1.0"
paths:
...
  /v1/attempt/create:
    post:
      consumes:
      - application/json
      description: 记录一次学习练习尝试
      parameters:
      - description: 学习记录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AttemptCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建学习记录
      tags:
      - 学习导航
  /v1/attempt/list:
    post:
      consumes:
      - application/json
      description: 分页查询当前用户的学习记录
      parameters:
      - description: 查询条件
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AttemptQuery'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取学习记录列表
      tags:
      - 学习导航
  /v1/attempt/recent:
    post:
      consumes:
      - application/json
      description: 获取指定目标最近的学习记录
      parameters:
      - description: 目标ID和数量限制
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取最近学习记录
      tags:
      - 学习导航
  /v1/attempt/stats:
    post:
      consumes:
      - application/json
      description: 获取指定目标的学习统计数据
      parameters:
      - description: 目标ID和可选章节ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取学习统计
      tags:
      - 学习导航
...
  /v1/download/{jwt}:
    get:
      description: 通过JWT令牌下载文档文件
      parameters:
      - description: JWT令牌
        in: path
        name: jwt
        required: true
        type: string
      - description: 文件名
        in: query
        name: filename
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
      summary: 下载文档
      tags:
      - 媒体
...
  /v1/upload/document:
    post:
      consumes:
      - multipart/form-data
      description: 上传PDF等文档文件（表单字段名为 file[0]）
      parameters:
      - description: 文档文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 上传失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 上传文档
      tags:
      - 媒体
  /v1/upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 上传失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 上传图片
      tags:
      - 媒体
...
  /v1/view/cover/{hash}:
    get:
      description: 根据文档哈希值获取文档封面图
      parameters:
      - description: 文档哈希
        in: path
        name: hash
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: 封面图片
      summary: 查看文档封面
      tags:
      - 媒体
  /v1/view/page/{hash}/{page}:
    get:
      description: 根据文档哈希和页码获取文档页面内容
      parameters:
      - description: 文档哈希
        in: path
        name: hash
        required: true
        type: string
      - description: 页面路径
        in: path
        name: page
        required: true
        type: string
      produces:
      - image/svg+xml
      responses:
        "200":
          description: 页面内容
      summary: 查看文档页面
      tags:
      - 媒体
...
securityDefinitions:
  BearerAuth:
    description: '输入 "Bearer {token}" 格式的 JWT，例如: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"